+++ two_d_game_ai/view/world_renderer.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import itertools
 import logging
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING
@@ -9,6 +10,7 @@ from typing import TYPE_CHECKING
 from pygame import Color, Rect, Surface, Vector2
 
 from two_d_game_ai.entities.bot import Bot
+from two_d_game_ai.entities.movement_block import MovementBlock
 from two_d_game_ai.view import colors
 from two_d_game_ai.view.bot_renderer import BotRenderer
 from two_d_game_ai.view.movement_block_renderer import MovementBlockRenderer
@@ -24,6 +26,7 @@ from two_d_game_ai.world.grid import Grid
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
+    from two_d_game_ai.view.generic_entity_renderer import GenericEntityRenderer
     from two_d_game_ai.world.world import World
 
 logger = logging.getLogger(__name__)
@@ -44,16 +47,15 @@ class WorldRenderer:
 
     surface: Surface = field(init=False)
     """pygame `Surface`."""
-    entity_renderers: dict[int, BotRenderer | MovementBlockRenderer] = field(
-        default_factory=dict
-    )
+    entity_renderers: dict[int, GenericEntityRenderer] = field(default_factory=dict)
     """Maps entity `id` to delegated renderer."""
-    selected: None | MovementBlockRenderer | BotRenderer = field(init=False)
+    selected_renderer: GenericEntityRenderer | None = field(init=False)
+    """Selected entity renderer."""
 
     def __post_init__(self) -> None:
         scaled_world_size = self.world.size * self.scale_factor
         self.surface = Surface((scaled_world_size, scaled_world_size))
-        self.selected = None
+        self.selected_renderer = None
         log_msg = f"WorldRenderer '{self.name}' initiated."
         logger.debug(log_msg)
 
@@ -76,7 +78,7 @@ class WorldRenderer:
         }
 
     @property
-    def clickables(self) -> set[BotRenderer | MovementBlockRenderer]:
+    def clickables(self) -> set[GenericEntityRenderer]:
         """Clickable elements."""
         return set(self.entity_renderers.values())
 
@@ -85,27 +87,27 @@ class WorldRenderer:
         self.ensure_renderers()
         # Drawn in order, bottom layer to top:
         self.surface.fill(colors.WORLD_FILL)
-        self._render_grid()
+        self.render_base_grid(self.world.grid)
+        self.render_untraversable_cells(self.world.grid)
         for b in self.bot_renderers:
             b.draw(debug_render_mode=debug_render_mode)
         for m in self.movement_block_renderers:
             m.draw()
-        self._render_axes()
+        self.render_axes()
 
     def ensure_renderers(self) -> None:
         """Update the set of entity renderers."""
-        for m in self.world.movement_blocks:
-            if m.id not in self.entity_renderers:
-                self.entity_renderers[m.id] = MovementBlockRenderer(
-                    world_renderer=self, entity=m
-                )
-                log_msg = f"{m.name} renderer added."
-                logger.debug(log_msg)
-
-        for b in self.world.bots:
-            if b.id not in self.entity_renderers:
-                self.entity_renderers[b.id] = BotRenderer(world_renderer=self, entity=b)
-                log_msg = f"{b.name} renderer added."
+        for e in self.world.entities:
+            if e.id not in self.entity_renderers:
+                if isinstance(e, MovementBlock):
+                    self.entity_renderers[e.id] = MovementBlockRenderer(
+                        world_renderer=self, entity=e
+                    )
+                elif isinstance(e, Bot):
+                    self.entity_renderers[e.id] = BotRenderer(
+                        world_renderer=self, entity=e
+                    )
+                log_msg = f"{e.name} renderer added."
                 logger.debug(log_msg)
 
     def to_local(self, world_pos: Vector2) -> Vector2:
@@ -147,48 +149,38 @@ class WorldRenderer:
         pos -= Vector2(offset, offset)
         return pos.elementwise() * Vector2(1, -1)
 
-    def _render_grid(self) -> None:
-        """Draw the `Grid`."""
-        grid_size = self.world.grid.size
-        cell_size = self.world.size / grid_size
+    def render_base_grid(self, grid: Grid) -> None:
+        """Draw the `Grid` nodes."""
+        cell_size = self.world.grid_resolution
+        cell_offsets = [
+            cell_size * i - self.world.magnitude + cell_size / 2
+            for i in range(grid.size)
+        ]
 
-        for cell_index in range(grid_size + 1):
-            cell_offset = cell_size * (cell_index - grid_size / 2)
-            # horizontal grid line
-            self.draw_line(
+        for x, y in itertools.product(cell_offsets, cell_offsets):
+            self.draw_circle(
                 color=colors.WORLD_GRID_LINE,
-                start_pos=Vector2(-self.world.magnitude, cell_offset),
-                end_pos=Vector2(self.world.magnitude, cell_offset),
-                width=1,
-                anti_alias=False,
-            )
-            # vertical grid line
-            self.draw_line(
-                color=colors.WORLD_GRID_LINE,
-                start_pos=Vector2(cell_offset, -self.world.magnitude),
-                end_pos=Vector2(cell_offset, self.world.magnitude),
-                width=1,
-                anti_alias=False,
+                center=Vector2(x, y),
+                radius=1,
+                scale_radius=False,
             )
 
-        oversize_px = 2
-        for cell_ref in self.world.grid.untraversable_cells:
+    def render_untraversable_cells(self, grid: Grid) -> None:
+        """Draw the blocked cells."""
+        cell_size = self.world.grid_resolution
+        for cell_ref in grid.untraversable_cells:
             grid_rect = Rect(
                 (cell_ref.x * cell_size, cell_ref.y * cell_size), (cell_size, cell_size)
             )
-            # draw slightly oversize to hide gridlines
-            oversize_grid_rect = grid_rect.move(
-                -int(oversize_px / self.scale_factor),
-                -int(oversize_px / self.scale_factor),
-            )
-            oversize_grid_rect.width = oversize_grid_rect.height = int(
-                grid_rect.width + 2 * oversize_px / self.scale_factor
-            )
+            # draw slightly oversize to avoid antialiasing issues
+            grid_rect = grid_rect.move(-0.5, -0.5)
+            grid_rect.width += 1
+            grid_rect.height += 1
             self.draw_rect(
-                color=colors.MOVEMENT_BLOCK_FILL, rect=Rect(oversize_grid_rect), width=0
+                color=colors.MOVEMENT_BLOCK_FILL, rect=Rect(grid_rect), width=0
             )
 
-    def _render_axes(self) -> None:
+    def render_axes(self) -> None:
         """Draw axes."""
         self.draw_line(
             color=colors.WORLD_AXES_LINE,
@@ -203,11 +195,11 @@ class WorldRenderer:
 
     def handle_mouse_select(self, local_pos: Vector2) -> None:
         """TO DO."""
-        self.selected = self.entity_at_pos(local_pos)
+        self.selected_renderer = self.entity_renderer_at_pos(local_pos)
         for renderer in self.clickables:
-            renderer.is_selected = renderer == self.selected
+            renderer.is_selected = renderer == self.selected_renderer
 
-    def entity_at_pos(self, pos: Vector2) -> MovementBlockRenderer | BotRenderer | None:
+    def entity_renderer_at_pos(self, pos: Vector2) -> GenericEntityRenderer | None:
         """Return the EntityRenderer at position, or None."""
         for renderer in self.clickables:
             if renderer.is_clicked(pos):
@@ -218,15 +210,15 @@ class WorldRenderer:
 
     def handle_mouse_set_destination(self, local_pos: Vector2) -> Vector2 | None:
         """Attempt to set destination, if applicable to current selection."""
-        if not isinstance(self.selected, BotRenderer):
-            return None
-        if not isinstance(self.selected.entity, Bot):
+        if not isinstance(self.selected_renderer, BotRenderer):
+            return None  # Only `Bot`s support destination setting
+        if not isinstance(self.selected_renderer.entity, Bot):
             raise TypeError
         world_pos = self.to_world(local_pos)
         cell = Grid.cell_from_world_pos(world=self.world, pos=world_pos)
         if not self.world.grid.is_traversable(cell):
             return None
-        self.selected.entity.destination = world_pos
+        self.selected_renderer.entity.destination = world_pos
         return world_pos
 
     def draw_line(